Code improvements, moved nmap code into modules/nmap.py
