+++ Lancer/config.py
@@ -1,0 +1,3 @@
+import argparse
+
+args = argparse.Namespace
