+++ requirements.txt
@@ -1,1 +1,2 @@
+pytest==5.1.0
 cpe_utils==0.1.0
